Replace use of gfloat by gdouble, since this is what Gtk+ 1.3 uses for  adjustments.
git-svn-id: svn+ssh://svn.eu/Dev/importfromcvs/trunk@71889 936e1b1b-40f2-da11-902a-00137254ae57

+++ widgets/cschtml/cschtml-keybinding.c
@@ -46,12 +46,12 @@ scroll_by_amount (CscHTML *html,
 {
 	GtkLayout *layout;
 	GtkAdjustment *adj;
-	gfloat new_value;
+	gdouble new_value;
 
 	layout = GTK_LAYOUT (html);
 	adj = layout->vadjustment;
 
-	new_value = adj->value + (gfloat) amount;
+	new_value = adj->value + (gdouble) amount;
 	if (new_value < adj->lower)
 		new_value = adj->lower;
 	else if (new_value > adj->upper)
